+++ packages/web3-eth/test/unit/web3_eth_methods_with_parameters.test.ts
@@ -16,6 +16,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { ethRpcMethods } from 'web3-rpc-methods';
 
+import { TransactionInfoAPI } from 'web3-types';
 import Web3Eth from '../../src/index';
 import * as rpcMethodWrappers from '../../src/rpc_method_wrappers';
 import {
@@ -52,7 +53,7 @@ import { testData as createAccessListTestData } from './rpc_method_wrappers/fixt
 jest.mock('web3-rpc-methods');
 jest.mock('../../src/rpc_method_wrappers');
 // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-jest.spyOn(rpcMethodWrappers, 'getTransaction').mockResolvedValue(tx);
+jest.spyOn(rpcMethodWrappers, 'getTransaction').mockResolvedValue(tx as TransactionInfoAPI);
 jest.spyOn(rpcMethodWrappers, 'getTransactionReceipt').mockResolvedValue(txReceipt);
 
 describe('web3_eth_methods_with_parameters', () => {